% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity_plot.R
\name{alpha_diversity_plot}
\alias{alpha_diversity_plot}
\title{alpha_diversity_plot}
\usage{
alpha_diversity_plot(phyloseq, feature, measures, p_test = "wilcox")
}
\arguments{
\item{phyloseq}{A phyloseq object contain otu table, taxonomy table, sample metadata and
phylogenetic tree.}

\item{feature}{The column name of the feature you want to select from metadata, e.g. "Phenotype".}

\item{measures}{The measures to calculate alpha diversity, measures should be one of "Observed",
"Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher".}

\item{p_test}{The p-value to test alpha diversity. p_test should be either "wilcox" or "kruskal".}
}
\description{
This is a function for plotting alpha diversity.
}
\examples{
alpha_diversity_plot(Shaoyifu_phyloseq, feature = "diagnosis",
                     measures = "Chao1", p_test = "kruskal")
}
