% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_otu_table.R
\name{construct_otu_table}
\alias{construct_otu_table}
\title{construct_otu_table}
\usage{
construct_otu_table(phyloseq, level = "all")
}
\arguments{
\item{phyloseq}{A phyloseq object contain otu table, taxonomy table, sample metadata and
phylogenetic tree.}

\item{level}{The coloumn name of the level wanted to select. Default is "all". If "all" then retain
all taxonomy level and seperate by "; ", else ONLY retain the given taxonomy level,
drop everything else. Level name should be one of "all", "Kingdom", "Phylum", "Class",
"Order", "Family", "Genus", "Species".}
}
\description{
construct_otu_table can construct a OTU table with a phyloseq object.
}
\examples{
construct_otu_table(Shaoyifu_phyloseq, level = "Order") \%>\% .[,1:5]
}
