% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_reads_track.R
\name{dada2_reads_track}
\alias{dada2_reads_track}
\title{dada2_reads_track}
\usage{
dada2_reads_track(track, single_end = FALSE)
}
\arguments{
\item{track}{The reads track data frame from Xbiome 16S pipeline dada2 workflow result.}

\item{single_end}{Default is FALSE. If single_end == TRUE, means the sequence files are single end,
the x-axis will contain 'input', 'filtered', 'dereplicated', 'nonchim'. If
single_end == FALSE, means the sequence files are paired end, the x-axis will
contain 'input', 'filtered', 'denoisedF', 'denoisedR', 'merged', 'nonchim'.}
}
\description{
dada2_reads_track can track the reads count in a dada2 workflow result which created by Xbiome 16S
pipeline. Xbiome 16S pipeline dada2 workflow will generate a list that contain sequence table,
taxonomy table and reads track data frame. Input the reads track data frame and read type, this
function can draw a line plot of reads track of every sample. X-axis will be every stage in dada2
workflow, Y-axis will be the reads counts.
}
\examples{
dada2_reads_track(Shaoyifu_dada2_result$reads_track, single_end = FALSE)
}
