% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2fc.R
\name{log2fc}
\alias{log2fc}
\title{log2fc}
\usage{
log2fc(phyloseq, feature, level = NA, p_value = 0.01, size = 2,
  save_res = FALSE)
}
\arguments{
\item{phyloseq}{A phyloseq object contain otu table, taxonomy table, sample metadata and
phylogenetic tree.}

\item{feature}{The column name of the feature you want to select from metadata, e.g. "Phenotype".}

\item{level}{Which taxonomy level to calculate fold change. Default is NA. If level is given, will
use construct_otu_table function to construct OTU table, and use DESeq to calculate
fold change.}

\item{p_value}{The cut off P value for the fold change. Default is 0.01.}

\item{save_res}{Default is FALSE. If TRUE, will save original result DESeq2_result.rds to current 
working directory.}
}
\description{
This is a function for plotting log2 fold change.
}
\examples{
log2fc(Shaoyifu_phyloseq, feature = "diagnosis", level = "Genus")
}
