% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{plot_cor}
\usage{
plot_cor(cor_tab, x, y, method = "pearson")
}
\arguments{
\item{cor_tab}{A data frame that contains at least two columns to calculate
correlation.}

\item{x}{Colname of cor_tab. x and y must have the same length.}

\item{y}{Colname of cor_tab. x and y must have the same length.}

\item{method}{A character string indicating which correlation coefficient is
to be used. One of "pearson", "kendall", or "spearman", can be abbreviated.
Default is "pearson".}
}
\description{
plot_cor can plot correlation using a input table.
}
