% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sparsity.R
\name{plot_sparsity}
\alias{plot_sparsity}
\title{plot_sparsity}
\usage{
plot_sparsity(otu_table, binwidth = NA)
}
\arguments{
\item{otu_table}{An OTU table in phyloseq object format. Rownames are 
sample ID, colnames are taxa.}

\item{binwidth}{The width of the bins. Can be specified as a numeric value, 
or a function that calculates width from x. The default is to use bins bins 
that cover the range of the data. You should always override this value, 
exploring multiple widths to find the best to illustrate the stories in your 
data.}
}
\description{
plot_sparsity can show the sparsity of an OTU table. It will calculate the 
prevalence of each OTU in all samples. X-axis is an OTU's prevalence. Y-axis 
is OTU count.
}
\examples{
plot_sparsity(dada2_res$seq_tab, 5)
}
