% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_diversity.R
\name{plot_beta_diversity}
\alias{plot_beta_diversity}
\title{plot_beta_diversity}
\usage{
plot_beta_diversity(phyloseq, feature, feature2 = NA, method)
}
\arguments{
\item{phyloseq}{A phyloseq object contain otu table, taxonomy table, sample metadata and
phylogenetic tree.}

\item{feature}{The column name of the feature you want to select from metadata, e.g. "Phenotype".}

\item{feature2}{The column name of another feature you want to select from metadata, which will
show in different shape, e.g. "Gender". Default is NA.}

\item{method}{The method to calculate beta diversity. Method should be one of "bray", "jaccard",
"unifrac", "wunifrac".}
}
\description{
This is a function for plotting beta diversity.
}
\examples{
plot_beta_diversity(demo_phyloseq_object, feature = "diagnosis", method = "bray")
}
