% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_bar.R
\name{plot_stacked_bar}
\alias{plot_stacked_bar}
\title{plot_stacked_bar}
\usage{
plot_stacked_bar(phyloseq, feature, level, x_size = 8,
  legend_position = "top", legend_size = 10)
}
\arguments{
\item{phyloseq}{A phyloseq object contain otu table, taxonomy table, sample metadata and
phylogenetic tree.}

\item{feature}{The feature that shows in x-axis text with different colors.}

\item{level}{Which taxonomy level to plot.}

\item{x_size}{The front size of x-axis text. Default is 8.}

\item{legend_position}{Legend position. Default is "top". legend_position should be one of "none",
"left", "right", "bottom", "top".}

\item{legend_size}{Lengend size. Default is 10.}
}
\description{
This is a function for stacked bar plot.
}
\examples{
plot_stacked_bar(demo_phyloseq_object, feature = "diagnosis", level = "Order")
}
