% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_bar.R
\name{plot_stacked_bar}
\alias{plot_stacked_bar}
\title{plot_stacked_bar}
\usage{
plot_stacked_bar(phyloseq = NULL, otu_table = NULL, metadata = NULL,
  level = NA, feature = NA, order = NULL, relative_abundance = TRUE,
  x_size = 8, legend_position = "top", legend_size = 10)
}
\arguments{
\item{phyloseq}{A phyloseq object contain otu table, taxonomy table, sample
metadata and phylogenetic tree. Set it to NULL if using an OTU table and
metadata to draw the stacked barplot.}

\item{otu_table}{An OTU table. Taxa are colnames, SampleID are rownames
(just like phyloseq). Set it to NULL if using phyloseq.}

\item{metadata}{A metadata for the OTU table. SampleID are rownames (just
like phyloseq). Set it to NULL if using phyloseq.}

\item{level}{A taxonomy level to plot stacked bar. Default is NA. Required
when using phyloseq object.}

\item{feature}{The feature that shows in x-axis text with different colors.
Default is NA, x-axis will show in black.}

\item{order}{A vector of arranged SampleID in wanted order. Default is NULL.
If NULL, plot_stacked_bar will automatically arrange samples by the most
abundance taxonomy in decreasing order.}

\item{relative_abundance}{Turn plot into relative abundance or not. Default
is TRUE.}

\item{x_size}{The front size of x-axis text. Default is 8.}

\item{legend_position}{Legend position. Default is "top". legend_position
should be one of "none", "left", "right", "bottom", "top".}

\item{legend_size}{Lengend size. Default is 10.}
}
\description{
This is a function for stacked barplot.
}
\examples{
plot_stacked_bar(demo_phyloseq_object, level = "Order")
}
