% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar_plot.R
\name{stacked_bar_plot}
\alias{stacked_bar_plot}
\title{Stacked bar plot of phylogenetic composition -------------------------------------------------------
This is a function for stacked bar plot.}
\usage{
stacked_bar_plot(phyloseq, feature, level)
}
\arguments{
\item{phyloseq}{A phyloseq object contain otu table, taxonomy table, sample metadata and
phylogenetic tree.}

\item{feature}{The feature that shows in x-axis text with different colors.}

\item{level}{Which taxonomy level to plot.}
}
\description{
Stacked bar plot of phylogenetic composition -------------------------------------------------------
This is a function for stacked bar plot.
}
\examples{
stacked_bar_plot(Shaoyifu_phyloseq, feature = "diagnosis", level = "Order")
}
